refactor(octokit-request-parser): pull the request parsing logic out of the logging plugin into a reusable module
